code_interpreter_tool/src/requirements.txt
bedrock-agentcore
strands-agents>=0.1.8
strands-agents-tools>=0.1.6
langchain-aws
langchain
pip
uv
boto3
bedrock-agentcore-starter-toolkit